Build script: list dependencies explicitly
With the current script, the jars in the classpaths are ordered differently
depending on the gradle version used (4+ incompatible with <=3), resulting in
gradle picking the incorrect version of classes overriden in gwt with the
super-source tag. Declaring the dependencies explicitly results in a finer
control of the classpath, working-around this problem.